✨ feat: jdbcTemplate를 사용한 insert 로그 기능 구현
- 페이지 방문, 검색 기록, 클릭 로그, 리뷰 조회 로그

🔗 Resolves: #158
