+++ .github/workflows/deploy-log.yml
@@ -3,13 +3,14 @@ name: Deploy to AWS EB
 # main 브랜치에 PR 병합되었을 때
 on:
   pull_request:
-    branches:
-      - develop
-    types:
-      - closed
+#    branches:
+#      - develop
+#    types:
+#      - closed
     # 해당 파일 또는 폴더 하위 파일들이 변경되었을 때에만!!!
     paths:
       - backend/JiShop/src/**
+      - backend/log/src/**
       - build.gradle
       - .github/workflows/deploy-dev.yml
       - .ebextensions/**
