--- backend/log/src/main/java/com/example/log/LogApplication.java
@@ -1,13 +1,0 @@
-package com.example.log;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LogApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LogApplication.class, args);
-    }
-
-}
