--- backend/log/src/main/java/com/example/log/common/exception/ErrorType.java
@@ -1,15 +1,0 @@
-package com.example.log.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum ErrorType {
-
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "사용자를 찾을 수 없습니다.");
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
