--- backend/log/src/main/java/com/example/log/common/exception/ExceptionResponse.java
@@ -1,8 +1,0 @@
-package com.example.log.common.exception;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class ExceptionResponse {
-    private String message;
-}
