--- backend/log/src/main/java/com/example/log/common/util/exampleUtil.java
@@ -1,4 +1,0 @@
-package com.example.log.common.util;
-
-public class exampleUtil {
-}
