--- backend/log/src/main/java/com/example/log/config/AsyncConfig.java
@@ -1,24 +1,0 @@
-package com.example.log.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@EnableAsync
-@Configuration
-public class AsyncConfig {
-    
-    @Bean(name = "logTaskExecutor")
-    public Executor logTaskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(5);
-        executor.setMaxPoolSize(10);
-        executor.setQueueCapacity(50);
-        executor.setThreadNamePrefix("logExecutor-");
-        executor.initialize();
-        return executor;
-    }
-}
