--- backend/log/src/main/java/com/example/log/config/webConfig.java
@@ -1,9 +1,0 @@
-package com.example.log.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class webConfig implements WebMvcConfigurer {
-
-}
