--- backend/log/src/main/java/com/example/log/controller/LogController.java
@@ -1,9 +1,0 @@
-package com.example.log.controller;
-
-import com.example.log.dto.LogRequest;
-import org.springframework.http.ResponseEntity;
-
-public interface LogController {
-
-    public ResponseEntity addLog(LogRequest logRequest);
-}
