--- backend/log/src/main/java/com/example/log/dto/LogRequest.java
@@ -1,14 +1,0 @@
-package com.example.log.dto;
-
-import com.example.log.domain.Log;
-import com.example.log.domain.enums.UserActionType;
-
-public record LogRequest(
-        Long userId,
-        String message,
-        UserActionType userActionType
-) {
-    public Log toEntity() {
-        return new Log(userId, message, userActionType);
-    }
-}
