--- backend/log/src/main/java/com/example/log/repository/LogRepository.java
@@ -1,7 +1,0 @@
-package com.example.log.repository;
-
-import com.example.log.domain.Log;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LogRepository extends JpaRepository<Log, Long> {
-}
