--- backend/log/src/main/java/com/example/log/service/LogService.java
@@ -1,7 +1,0 @@
-package com.example.log.service;
-
-import com.example.log.dto.LogRequest;
-
-public interface LogService {
-    public void addLog(LogRequest logRequest);
-}
