+++ backend/log/src/main/java/com/log/LogApplication.java
@@ -1,0 +1,13 @@
+package com.log;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LogApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LogApplication.class, args);
+    }
+
+}
