+++ backend/log/src/main/java/com/log/common/exception/DomainException.java
@@ -1,0 +1,21 @@
+package com.log.common.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class DomainException extends RuntimeException {
+
+    private final String message;
+    private final HttpStatus status;
+
+    public DomainException(ErrorType errorType) {
+        this.message = errorType.getMessage();
+        this.status = errorType.getHttpStatus();
+    }
+
+    public ExceptionResponse toResponse() {
+        return new ExceptionResponse(message);
+    }
+
+}
