+++ backend/log/src/main/java/com/log/common/exception/ExceptionResponse.java
@@ -1,0 +1,8 @@
+package com.log.common.exception;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class ExceptionResponse {
+    private String message;
+}
