+++ backend/log/src/main/java/com/log/common/util/exampleUtil.java
@@ -1,0 +1,4 @@
+package com.log.common.util;
+
+public class exampleUtil {
+}
