+++ backend/log/src/main/java/com/log/config/webConfig.java
@@ -1,0 +1,9 @@
+package com.log.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class webConfig implements WebMvcConfigurer {
+
+}
