+++ backend/log/src/main/java/com/log/dto/PageViewRequest.java
@@ -1,0 +1,11 @@
+package com.log.dto;
+
+import java.time.LocalDateTime;
+
+public record PageViewRequest(
+        Long userId,
+        String pageUrl,
+        LocalDateTime visitTime,
+        LocalDateTime exitTime,
+        Integer durationSeconds
+) { }
