+++ backend/log/src/main/java/com/log/dto/ProductClickRequest.java
@@ -1,0 +1,11 @@
+package com.log.dto;
+
+import java.time.LocalDateTime;
+
+public record ProductClickRequest(
+        Long userId,
+        Long productId,
+        String productName,
+        LocalDateTime clickTime
+) {
+}
