+++ backend/log/src/main/java/com/log/dto/ReviewRequest.java
@@ -1,0 +1,12 @@
+package com.log.dto;
+
+import java.time.LocalDateTime;
+
+public record ReviewRequest(
+        Long userId,
+        Long productId,
+        String productName,
+        Long reviewId,
+        LocalDateTime viewTime
+) {
+}
