+++ backend/log/src/main/java/com/log/dto/SearchRequest.java
@@ -1,0 +1,10 @@
+package com.log.dto;
+
+import java.time.LocalDateTime;
+
+public record SearchRequest(
+        Long userId,
+        String searchKeyword,
+        LocalDateTime searchTime
+) {
+}
