--- backend/log/src/test/java/com/example/log/LogApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.log;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LogApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
