+++ backend/log/src/test/java/com/log/LogApplicationTests.java
@@ -1,0 +1,13 @@
+package com.log;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LogApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
